Rename test file
To follow Ruby gems convention.
